K: moved note routes to a single file
