+++ ktor/src/main/kotlin/server/Main.kt
@@ -7,12 +7,13 @@ import io.ktor.routing.Routing
 import io.ktor.serialization.json
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
+import org.kodein.di.bind
 import org.kodein.di.ktor.di
+import org.kodein.di.singleton
 import server.composition.databaseModule
-import server.routes.addNotePostRoute
-import server.routes.notesGetRoute
-import server.routes.rootGetRoute
-import server.routes.updateNotePatchRoute
+import server.routes.notes.NotesCallHelper
+import server.routes.notes.notesRoute
+import server.routes.rootRoute
 
 fun main() {
     embeddedServer(
@@ -29,13 +30,12 @@ private fun getPort() = System.getenv("PORT")?.toIntOrNull() ?: 9000
 fun Application.module() {
     di {
         import(databaseModule)
+        bind() from singleton { NotesCallHelper() }
     }
 
     install(Routing) {
-        rootGetRoute()
-        notesGetRoute()
-        updateNotePatchRoute()
-        addNotePostRoute()
+        rootRoute()
+        notesRoute()
     }
     install(ContentNegotiation) {
         json()
