+++ ktor/src/main/kotlin/server/routes/rootGetRoute.kt
@@ -6,7 +6,7 @@ import io.ktor.response.respondText
 import io.ktor.routing.Routing
 import io.ktor.routing.get
 
-fun Routing.rootGetRoute() {
+fun Routing.rootRoute() {
     get("/") {
         call.respondText("Ktor server", ContentType.Text.Html)
     }
