+++ ktor/src/main/kotlin/server/storage/DatabaseCoroutineQuery.kt
@@ -1,4 +1,4 @@
-package server.routes.helpers
+package server.storage
 
 import kotlinx.coroutines.Dispatchers
 import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
